break out rtree implementation in to separate packages
